Added HDA for procedurally loading entire ldraw files, as an alternative to the shelf script approach.
Factored out some functions in ldraw.py so they can be used in both
approaches.

Realized that the post-compensation in xform is just the inverted
initial matrix, so bake that in before setting node/point xforms.

Fetch the color_lib() at the module level. Might not be the best idea
but easy for now.

Lacking some features in the HDA and the build_model_points path in
general. We should still expose all the LDraw Part options (gaps etc.)
and store material paths on prims. I probably missed some other bits
too.